Revert "gtimezone: Cache timezones based on the identifier they were created by"
This reverts commit 851241f19a3fd9ec693b3dd8f37a84c7f970984a.

That commit avoids a performance regression but introduces a behavior regression:
changes to /etc/localtime have no effect for the remaining of the application's
runtime.

With the optimization introduced by the previous commit, we can pass NULL to
g_time_zone_new() repeatedly with no performance drawback, so we no longer have
to workaround this case.